// src/models/Content.ts
export interface Content {
  id: number;
  type: string;
  state: string;
  title: string;
  description: string;
  cartelImage: string;
  heroImage: string;
  category: string;
  createdBy: number;
  lastUpdated: Date;
  createdAt: Date;
  duration: number;
  trailer: string;
  DVDVersion: string;
}